Refactor: Item service broken into ItemHelper and ItemMover services.
